Implementation of functions
Changed the prototypes of most functions to point to a general linked_list_node rather than just the head node. This way, most operations can be called relative to a node in the linked list. This might be a problem if the user frees a node that is pointed to by another node, however it might work if all the nodes are correctly NULL-afied upon freeing.