Update simplemain.c
Specific test added for using the add and remove operations when node given is not the head of the list. (CAUTION: this should never be used unless you absolutely know what you are doing. Removing nodes that are pointed by pointers in the main will result in a segmentation fault. This is very easy to do if, for example, the head gets removes with remove_head_node from a node that is not the head. If there was a pointer in main to head - ex. head1 - this node will now be deallocated, resulting in a segmentation fault.) One possible solution is to always turn the node sent into the head, and then the user just has to use that as the new head instead. This will still result in problems in case there are pointers to specific nodes in the list that are not the head, however in that case the user is more at fault for freeing a memory position pointed by an node they were using (in fact this can also happen in the singly linked list). One solution is to hide the get_next_node/get_previous_node/next_node/previous_node from the user and instead add a search function to the implementation for the user to specify a place to remove or get a value. However, this implies sending a comparator function by the user into the implementation which is an interesting solution, but will complicate everything. If the user can correctly create a comparator function and send it as input to a function, it is likely that this user will not make seg faults.